Directory based extention loading
