+++ main.py
@@ -1,12 +1,13 @@
 import logging
 from logging import config
 
-from os import getenv
+from os import getenv, listdir
 from dotenv import load_dotenv
 load_dotenv()
 
 import discord
 from discord.ext import commands
+from discord import ExtensionAlreadyLoaded
 
 # ---------------------> Logging setup
 
@@ -19,13 +20,18 @@ intents = discord.Intents.all()
 bot = commands.Bot(command_prefix = '$', intents = intents)
 
 @bot.event
-async def on_ready():
-    log.info(f'Succesful login as {bot.user}')
+async def on_ready() -> None:
+    log.info(f'Succesful login as {bot.user}')      
 
 # ---------------------> Main
 
 if __name__ == '__main__':
-
-    from cogs.example_cog import Example # TODO anything but this
-    bot.add_cog(Example(bot))
+    for ext in ['cogs.' + file[:-3] for file in listdir('./cogs') if file.endswith('.py')]:
+        try:
+            bot.load_extension(ext)
+        except ExtensionAlreadyLoaded as err:
+            log.warning(err)
+        except Exception as err:
+            log.error(err)
+        
     bot.run(getenv('DISCORD_TOKEN'))
